refactor(core): promote PokeAPI sprite URL helpers to Pkmds.Core
Moves GetPokeApiHomeSpriteUrl, GetPokeApiVersionSpriteUrl and their
form-mapping data tables (Mega/regional/gender/Alcremie/Vivillon/Unown/etc.)
out of Pkmds.Rcl/ImageHelper.cs into a new Pkmds.Core/Utilities/PokeApiSpriteUrls
class so AOT-clean consumers — the macOS and iOS Quick Look extensions —
can share the full lookup instead of falling back to base species sprites.

The macOS/iOS HtmlRenderer.cs files now call the shared helper with
species/form/formArg/gender/isShiny, and the local BuildHomeSpriteUrl
fallback is gone. Bundled-sprite path helpers (`_content/Pkmds.Rcl/...`)
stay in Rcl since they're only valid inside the Blazor static-web-asset
host. Tracking the remaining audit candidates in #804.
