+++ Pkmds.Rcl/Components/Dialogs/VivillonEditorDialog.razor
@@ -38,7 +38,7 @@
                                   @onclick="@(() => selectedForm = formIdx)">
                             <img src="@(HostService.IsEmbedded || failedFormSprites.Contains(formIdx)
                                           ? ImageHelper.GetPokemonSpriteFilenameForForm(vivillonSpecies, Pokemon.Context, formIdx)
-                                          : ImageHelper.GetPokeApiHomeSpriteUrl(vivillonSpecies, formIdx, null, isPreviewShiny, Pokemon.Gender)
+                                          : PokeApiSpriteUrls.GetPokeApiHomeSpriteUrl(vivillonSpecies, formIdx, null, isPreviewShiny, Pokemon.Gender)
                                             ?? ImageHelper.GetPokemonSpriteFilenameForForm(vivillonSpecies, Pokemon.Context, formIdx))"
                                  alt="@forms[formIdx]"
                                  style="width:64px; height:48px; object-fit:contain; display:block;"
