+++ Pkmds.Rcl/Components/PokemonSlotComponent.razor
@@ -25,10 +25,10 @@
             var highResUrl = HostService.IsEmbedded
                 ? null
                 : AppState.SpriteStyle == SpriteStyle.Game
-                    ? ImageHelper.GetPokeApiVersionSpriteUrl(Pokemon.Species, Pokemon.Form,
+                    ? PokeApiSpriteUrls.GetPokeApiVersionSpriteUrl(Pokemon.Species, Pokemon.Form,
                         Pokemon.GetFormArgument(0), isShiny, Pokemon.Gender,
                         AppState.SaveFile?.Version ?? GameVersion.Any)
-                    : ImageHelper.GetPokeApiHomeSpriteUrl(Pokemon.Species, Pokemon.Form,
+                    : PokeApiSpriteUrls.GetPokeApiHomeSpriteUrl(Pokemon.Species, Pokemon.Form,
                         Pokemon.GetFormArgument(0), isShiny, Pokemon.Gender);
             <object class="pkm-sprite w-full h-auto object-contain block"
                     style="@(highResLoaded
